+++ src/auth0__jwks_cert_/index.d.ts
@@ -1,7 +1,0 @@
-import type { Ctx } from '@ctx-core/object'
-export declare function auth0__jwks_cert_(
-	ctx:Ctx
-):Promise<string>
-export {
-	auth0__jwks_cert_ as jwks_cert_,
-}