+++ src/auth0__jwks_cert__new/index.d.ts
@@ -1,0 +1,7 @@
+import type { Ctx } from '@ctx-core/object'
+export declare function auth0__jwks_cert__new(
+	ctx:Ctx
+):Promise<string>
+export {
+	auth0__jwks_cert__new as jwks_cert_,
+}