+++ src/auth0__jwks_cert__new/index.js
@@ -1,11 +1,11 @@
-import { auth0__jwks_x5c_ } from '../auth0__jwks_x5c_/index.js'
+import { auth0__jwks_x5c__new } from '../auth0__jwks_x5c__new/index.js'
 /**
  * @param {import('@ctx-core/object').Ctx}ctx
  * @return {Promise<string>}
  * @private
  */
-export async function auth0__jwks_cert_(ctx) {
-	const jwks_x5c = await auth0__jwks_x5c_(ctx)
+export async function auth0__jwks_cert__new(ctx) {
+	const jwks_x5c = await auth0__jwks_x5c__new(ctx)
 	const injwks_cert_fn = jwks_x5c[0]
 	const jwks_cert =
 		['-----BEGIN CERTIFICATE-----',
@@ -15,5 +15,6 @@ export async function auth0__jwks_cert_(ctx) {
 	return jwks_cert
 }
 export {
-	auth0__jwks_cert_ as jwks_cert_,
+	auth0__jwks_cert__new as auth0__jwks_cert_,
+	auth0__jwks_cert__new as jwks_cert_,
 }