+++ src/auth0__jwks_x5c_/index.d.ts
@@ -1,7 +1,0 @@
-import type { Ctx } from '@ctx-core/object'
-export declare function auth0__jwks_x5c_(
-	ctx:Ctx
-):Promise<string>
-export {
-	auth0__jwks_x5c_ as jwks_x5c_,
-}