+++ src/auth0__jwks_x5c__new/index.d.ts
@@ -1,0 +1,8 @@
+import type { Ctx } from '@ctx-core/object'
+export declare function auth0__jwks_x5c__new(
+	ctx:Ctx
+):Promise<string>
+export {
+	auth0__jwks_x5c__new as auth0__jwks_x5c_,
+	auth0__jwks_x5c__new as jwks_x5c_,
+}